Correcting format in imports, adding metadata option to mobile view, renaming IMetadata.ts to PaginatedMetadata.ts, moving subscribe to onInit method.
