--- ui/src/app/metadata/IMetadata.ts
@@ -1,25 +1,0 @@
-import {ApiJobCode} from "./job-codes/Jobcode"
-import {ApiNamedReference} from "./named-references/NamedReference"
-
-export interface IMetadata {
-  apiJobCodes?: ApiJobCode[],
-  apiNamedReferences?: ApiNamedReference[]
-}
-
-export class ApiMetadata implements IMetadata {
-
-  constructor(o?: IMetadata) {
-    if (o !== undefined) {
-      Object.assign(this, o)
-    }
-  }
-}
-
-export class PaginatedMetadata {
-  totalCount = 0
-  metadata: ApiJobCode[]|ApiNamedReference[] = []
-  constructor(metadata: ApiJobCode[]|ApiNamedReference[], totalCount: number) {
-    this.metadata = metadata
-    this.totalCount = totalCount
-  }
-}
