+++ ui/src/app/metadata/detail/metadata-list/metadata-list.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing"
 import { MetadataListComponent } from "./metadata-list.component"
 import { MetadataType } from "../../rsd-metadata.enum"
-import { PaginatedMetadata } from "../../IMetadata"
+import { PaginatedMetadata } from "../../PaginatedMetadata"
 import { ApiJobCode } from "../../job-codes/Jobcode"
 import { AuthServiceStub } from "@test/resource/mock-stubs";
 import { AuthService } from "../../../auth/auth-service";
