+++ ui/src/app/navigation/commoncontrols-mobile.component.html
@@ -16,6 +16,12 @@
       <span class="m-buttonText-x-text">Categories</span>
     </button>
   </li>
+
+  <li>
+    <button class="m-buttonText m-buttonText-onDark" routerLink="/metadata">
+      <span class="m-buttonText-x-text">Metadata</span>
+    </button>
+  </li>
 </ul>
 
 <div>
